Faraday no longer attempts to verify SSL by default.
Faraday used to attempt to verify SSL connections by default, which resulted
in "peer certificate won't be verified in this SSL session" warnings which
couldn't be suppressed; we got around this with a nasty hack: creating a
new Faraday middleware and duplicating the Tinder connection setup methods
to insert it.

The latest version of Faraday no longer attempts the verification by default,
which means we don't need our nasty hack any more.